Setting up pagination for sections on printed page.
The header for each section marked as "multi-page"
should now repeate if the section was split into multiple
pages.  This only works in Firefox (tested) and supposedly
IE (Not tested yet).  One issue is that if the first
section was needed to be pushed to the next page, the
header will show up by itself on the first page, then
again with the section on the next page.